requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-login==0.6.3
flask-cors==4.0.1
python-dotenv==1.0.1
eventlet==0.35.2
gunicorn==22.0.0

# Supabase
supabase==2.9.1
httpx==0.27.2
gotrue==2.9.1
postgrest==0.17.1
realtime==2.0.2
storage3==0.8.1
python-jose==3.3.0

# Auth
authlib==1.3.1
requests==2.31.0
bcrypt==4.1.3

# AI / ML
tensorflow==2.16.1
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.5.0

# Computer Vision
opencv-python-headless==4.10.0.84
mediapipe>=0.10.9
mtcnn==0.1.1
deepface==0.0.93
Pillow==10.3.0

# Audio Processing
librosa==0.10.2
soundfile==0.12.1
noisereduce==3.0.2
pyaudio==0.2.14

# Utilities
ffmpeg-python==0.2.0
cryptography==42.0.7
pyOpenSSL==24.1.0
python-multipart==0.0.9
aiofiles==23.2.1
